move logic to useDuplicateDashCard
